added code for parsing data and computing odometry and creating odometry graph

+++ main/include/encoder.h
@@ -30,4 +30,4 @@ void encoder_task(void);
 BaseType_t encoder_service(void);
 void enc_buf_to_text();
 void enc_time_to_buf(int32_t time_ms);
-void enc_data_to_buf(int32_t data);
+void enc_data_to_buf(float data);
